+++ zebra-chain/src/sprout/tests/arbitrary.rs
@@ -1,0 +1,57 @@
+use proptest::{arbitrary::any, array, prelude::*};
+
+use crate::{
+    amount::{Amount, NonNegative},
+    primitives::ZkSnarkProof,
+};
+
+use super::super::{commitment, note, tree, JoinSplit};
+
+impl<P: ZkSnarkProof + Arbitrary + 'static> Arbitrary for JoinSplit<P> {
+    type Parameters = ();
+
+    fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
+        (
+            any::<Amount<NonNegative>>(),
+            any::<Amount<NonNegative>>(),
+            any::<tree::NoteTreeRootHash>(),
+            array::uniform2(any::<note::Nullifier>()),
+            array::uniform2(any::<commitment::NoteCommitment>()),
+            array::uniform32(any::<u8>()),
+            array::uniform32(any::<u8>()),
+            array::uniform2(any::<note::MAC>()),
+            any::<P>(),
+            array::uniform2(any::<note::EncryptedCiphertext>()),
+        )
+            .prop_map(
+                |(
+                    vpub_old,
+                    vpub_new,
+                    anchor,
+                    nullifiers,
+                    commitments,
+                    ephemeral_key_bytes,
+                    random_seed,
+                    vmacs,
+                    zkproof,
+                    enc_ciphertexts,
+                )| {
+                    Self {
+                        vpub_old,
+                        vpub_new,
+                        anchor,
+                        nullifiers,
+                        commitments,
+                        ephemeral_key: x25519_dalek::PublicKey::from(ephemeral_key_bytes),
+                        random_seed,
+                        vmacs,
+                        zkproof,
+                        enc_ciphertexts,
+                    }
+                },
+            )
+            .boxed()
+    }
+
+    type Strategy = BoxedStrategy<Self>;
+}
